BackEnd/Django/modelImport/requirements.txt
absl-py==1.4.0
accelerate @ git+https://github.com/huggingface/accelerate.git@55691b14c21748319bc7004f09d2ea6019e71a25
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
diffusers==0.14.0
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
filelock==3.10.7
flatbuffers==23.5.9
fsspec==2023.4.0
ftfy==6.1.1
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio_client==0.1.3
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
jax==0.4.10
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
networkx==3.0
neural==0.1.0
neural-style==0.5.7
neuralstyle==0.0.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
Pillow==9.5.0
protobuf==4.22.3
psutil==5.9.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyre-extensions==0.0.23
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.3.4
wrapt==1.14.1
xformers==0.0.18
zipp==3.15.0
